+++ CHANGELOG.md
@@ -2,9 +2,26 @@
 
 The following is a history of the changes made to this project.
 
-## v2.8.2 *(June 29th, 2024)* :baby_chick:
+## v2.8.3 *(July 7th, 2024)* :chicken:
 
 ### Installers
+
+* Added the option to install the Fly Italy ADS-B feeder client.
+* Tested installation processes on Arbian Bookworm.
+* Tested installation processes on Arbian Jammy.
+* Unnecessary RTL-SDR kernel modules are once again blacklisted during decoder installs. 
+* ADS-B Receiver Project repository updates now work properly when executed by the script.
+* Instead of creating a new branch changes to tracked files are now stashed before updating.
+* The --development and -d parameters are now respected by the install scripts.
+* The --branch <branch_name> and -b <branch_name> parameters properly specify the branch to use.
+* Fixed issue noticed in the line downloading the duck DNS log during setup.
+* Refactoring and standards compliance changes were made to multiple bash scripts.
+
+### Portal
+
+* The proper collectd CPU temperature configuration is inserted for Raspberry PI 4 Model B devices.
+
+## v2.8.2 *(June 29th, 2024)* :baby_chick:
 
 * Added the option to install the ADS-B Exchange stats package after feeder installation.
 * Added the option to install the ADS-B Exchange web interface after feeder installation.
@@ -12,9 +29,6 @@ The following is a history of the changes made to this project.
 * Added support for Ubuntu Focal Fossa.
 * Added support for DietPi Bookworm.
 * Added support for DietPi Bullseye.
-
-### Portal
-
 * Fixed issue where navigation element toggling was not working in lite installs. 
 * Added eno1 and wlo1 to network options in the administration settings section of the portal.
 * The ADS-B Max Range temperature graph is now displayed in system performace the graphs collection.